Only use double-precision floating point
* Avoids `-Wdouble-promotion` in consumers

+++ src/AlignSettings.cpp
@@ -53,7 +53,7 @@ const CLI_v2::Option MinPercConcordance{
 R"({
     "names" : ["c", "min-concordance-perc"],
     "description" : "Minimum alignment concordance in percent.",
-    "type" : "float",
+    "type" : "double",
     "default" : 70
 })"};
 
@@ -61,7 +61,7 @@ const CLI_v2::Option MinPercIdentity{
 R"({
     "names" : ["x", "min-id-perc"],
     "description" : "Minimum sequence identity in percent.",
-    "type" : "float",
+    "type" : "double",
     "default" : 0
 })"};
 
@@ -69,7 +69,7 @@ const CLI_v2::Option MinPercIdentityGapComp{
 R"({
     "names" : ["y", "min-gap-comp-id-perc"],
     "description" : "Minimum gap compressed sequence identity in percent.",
-    "type" : "float",
+    "type" : "double",
     "default" : 0
 })"};
